% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_book.R
\name{my_book}
\alias{my_book}
\title{A test for rendering bookdown pdf}
\usage{
my_book(toc = TRUE, number_sections = TRUE, fig_caption = TRUE, ...,
  base_format = bookdown_test)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{...}{Other arguments to be passed to \code{base_format}.}

\item{base_format}{An output format function to be used as the base format.}
}
\description{
A test for rendering bookdown pdf
}

